+++ src/index.html
@@ -69,7 +69,6 @@
       <button type="submit" class="upload">Upload</button>
     </form>
 
-    <script src="scripts/destinationList.js"></script>
     <script src="main.js"></script>
   </body>
 </html>
